Add __ne__ magic method.
In Python 2, if you do not define a magic method, such as `__eq__`  for equality, or `__ne__` for inequality, it defaults to comparing the object IDs, which are random. That causes checking for inequality (`msg_1 != msg_2`) to work unexpectedly. (i.e., even if the messages are equal, it would return True.) In fact, the [Python 2 docs](https://docs.python.org/2/reference/datamodel.html#basic-customization) say that it is best to add them too.

In Python 3, `__ne__` is simply the inverse of `__eq__` if it is not defined.

Test adding the __ne__ magic method.