Align the nsecs field of the timestamps.
This change ensures that the nsecs field is always right aligned such
that the least significant digit of the nsecs field aligns with the
least significant digit of the secs field. This makes it easier to
inspect how far into the second the time has progressed.

The format specification explains the intent; always one space and
aligned to the right with a minimum width of 9 characters.