Update regression test for nsecs alignment.
The alignment of the nsecs field was changed in ros/genpy@edb9c5abf2b4fe7d716ae8fa72cceb24717b0035
(pull request ros/genpy#45), this caused the regression test for
the timestamp field to fail.

This commit updates the comparison value of the regression tests to reflect
the new alignment.